requirements.txt
django==3.2.25
pymysql==1.0.2
djangorestframework==3.15.1
drf-yasg==1.21.7
django-cors-headers